requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
django-rest-framework==0.1.0
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysql-connector-python==9.5.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0
